+++ src/main/java/com/example/bsc/dto/PaymentDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Getter
 @ToString
@@ -15,5 +17,5 @@ public class PaymentDTO {
 	/**
 	 * how many of this currency is added (positive)/substracted(negative) from balance
 	 */
-	int amount;
+	BigDecimal amount;
 }
